// backend/api/lib/angelone/client.ts
export interface MarketQuoteFull {
  exchange:       string;
  tradingSymbol:  string;
  symbolToken:    string;
  ltp:            number;
  open:           number;
  high:           number;
  low:            number;
  close:          number;
  lastTradeQty:   number;
  exchFeedTime:   string;
  exchTradeTime:  string;
  netChange:      number;
  percentChange:  number;
  avgPrice:       number;
  tradeVolume:    number;
  opnInterest:    number;
  upperCircuit:   string;
  lowerCircuit:   string;
  totBuyQuan:     number;
  totSellQuan:    number;
  '52WeekLow':    number;
  '52WeekHigh':   number;
  depth?: {
    buy:  Array<{ price: number; quantity: number; orders: number }>;
    sell: Array<{ price: number; quantity: number; orders: number }>;
  };
}